+++ src/Label/IntegerGenerator.php
@@ -16,9 +16,11 @@ namespace Bakame\Period\Visualizer\Label;
 use League\Period\Sequence;
 use function array_map;
 use function count;
+use function is_scalar;
+use function method_exists;
 use function range;
 
-final class IntegerType implements LabelGenerator
+final class IntegerGenerator implements LabelGenerator
 {
     /**
      * @var int
@@ -70,13 +72,25 @@ final class IntegerType implements LabelGenerator
     /**
      * {@inheritdoc}
      */
-    public function generateLabels(Sequence $sequence): array
+    public function generate(Sequence $sequence): array
     {
         $letters = [];
         if ($sequence->isEmpty()) {
             return $letters;
         }
 
-        return array_map('sprintf', range($this->int, $this->int + count($sequence) - 1));
+        return array_map([$this, 'format'], range($this->int, $this->int + count($sequence) - 1));
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function format($str): string
+    {
+        if (is_scalar($str) || method_exists($str, '__toString') || null === $str) {
+            return (string) $str;
+        }
+
+        return '';
     }
 }