+++ src/Label/RomanGenerator.php
@@ -16,8 +16,10 @@ namespace Bakame\Period\Visualizer\Label;
 use League\Period\Sequence;
 use function array_map;
 use function in_array;
+use function strtolower;
+use const FILTER_VALIDATE_INT;
 
-final class RomanType implements LabelGenerator
+final class RomanGenerator implements LabelGenerator
 {
     public const UPPER = 1;
     public const LOWER = 2;
@@ -31,7 +33,7 @@ final class RomanType implements LabelGenerator
     ];
 
     /**
-     * @var IntegerType
+     * @var IntegerGenerator
      */
     private $labelGenerator;
 
@@ -43,7 +45,7 @@ final class RomanType implements LabelGenerator
     /**
      * New instance.
      */
-    public function __construct(IntegerType $labelGenerator, int $lettercase = self::UPPER)
+    public function __construct(IntegerGenerator $labelGenerator, int $lettercase = self::UPPER)
     {
         $this->labelGenerator = $labelGenerator;
         $this->lettercase = $this->filter($lettercase);
@@ -118,14 +120,9 @@ final class RomanType implements LabelGenerator
     /**
      * {@inheritdoc}
      */
-    public function generateLabels(Sequence $sequence): array
+    public function generate(Sequence $sequence): array
     {
-        $retval = array_map([$this, 'convert'], $this->labelGenerator->generateLabels($sequence));
-        if (self::LOWER === $this->lettercase) {
-            return array_map('strtolower', $retval);
-        }
-
-        return $retval;
+        return array_map([$this, 'format'], $this->labelGenerator->generate($sequence));
     }
 
     /**
@@ -146,6 +143,23 @@ final class RomanType implements LabelGenerator
             }
         }
 
+        if (self::LOWER === $this->lettercase) {
+            return strtolower($retVal);
+        }
+
         return $retVal;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function format($str): string
+    {
+        $res = filter_var($str, FILTER_VALIDATE_INT, ['options' => ['min_range' => 1]]);
+        if (false === $res) {
+            return '';
+        }
+
+        return $this->convert($res);
     }
 }