+++ tests/Label/AffixGeneratorTest.php
@@ -13,18 +13,18 @@ declare(strict_types=1);
 
 namespace BakameTest\Period\Visualizer\Label;
 
-use Bakame\Period\Visualizer\Label\AffixType;
-use Bakame\Period\Visualizer\Label\IntegerType;
-use Bakame\Period\Visualizer\Label\LetterType;
-use Bakame\Period\Visualizer\Label\RomanType;
+use Bakame\Period\Visualizer\Label\AffixGenerator;
+use Bakame\Period\Visualizer\Label\IntegerGenerator;
+use Bakame\Period\Visualizer\Label\LetterGenerator;
+use Bakame\Period\Visualizer\Label\RomanGenerator;
 use League\Period\Period;
 use League\Period\Sequence;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @coversDefaultClass Bakame\Period\Visualizer\Label\AffixType;
+ * @coversDefaultClass \Bakame\Period\Visualizer\Label\AffixType;
  */
-final class AffixTypeTest extends TestCase
+final class AffixGeneratorTest extends TestCase
 {
     /**
      * @dataProvider providerLetter
@@ -36,8 +36,11 @@ final class AffixTypeTest extends TestCase
         string $suffix,
         array $expected
     ): void {
-        $generator = (new AffixType(new LetterType($letter)))->withPrefix($prefix)->withSuffix($suffix);
-        self::assertSame($expected, $generator->generateLabels($sequence));
+        $generator = new AffixGenerator(new LetterGenerator($letter), $prefix, $suffix);
+        self::assertSame($expected, $generator->generate($sequence));
+
+        $generator = (new AffixGenerator(new LetterGenerator($letter)))->withPrefix($prefix)->withSuffix($suffix);
+        self::assertSame($expected, $generator->generate($sequence));
     }
 
     public function providerLetter(): iterable
@@ -93,12 +96,18 @@ final class AffixTypeTest extends TestCase
 
     public function testGetter(): void
     {
-        $generator = new AffixType(new RomanType(new IntegerType(10)));
+        $generator = new AffixGenerator(new RomanGenerator(new IntegerGenerator(10)));
         self::assertSame('', $generator->getSuffix());
         self::assertSame('', $generator->getPrefix());
         $new = $generator->withPrefix('o')->withSuffix('');
         self::assertNotSame($new, $generator);
         self::assertSame('o', $new->getPrefix());
         self::assertSame('', $new->getSuffix());
+    }
+
+    public function testFormat(): void
+    {
+        $generator = new AffixGenerator(new RomanGenerator(new IntegerGenerator(10)), ':', '.');
+        self::assertSame(':.', $generator->format([]));
     }
 }