+++ tests/Label/RomanGeneratorTest.php
@@ -13,24 +13,24 @@ declare(strict_types=1);
 
 namespace BakameTest\Period\Visualizer\Label;
 
-use Bakame\Period\Visualizer\Label\IntegerType;
-use Bakame\Period\Visualizer\Label\RomanType;
+use Bakame\Period\Visualizer\Label\IntegerGenerator;
+use Bakame\Period\Visualizer\Label\RomanGenerator;
 use League\Period\Period;
 use League\Period\Sequence;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @coversDefaultClass Bakame\Period\Visualizer\Label\RomanType;
+ * @coversDefaultClass \Bakame\Period\Visualizer\Label\RomanType;
  */
-final class RomanTypeTest extends TestCase
+final class RomanGeneratorTest extends TestCase
 {
     /**
      * @dataProvider providerLetter
      */
     public function testGetLabels(Sequence $sequence, int $label, int $lettercase, array $expected): void
     {
-        $generator = new RomanType(new IntegerType($label), $lettercase);
-        self::assertSame($expected, $generator->generateLabels($sequence));
+        $generator = new RomanGenerator(new IntegerGenerator($label), $lettercase);
+        self::assertSame($expected, $generator->generate($sequence));
     }
 
     public function providerLetter(): iterable
@@ -39,7 +39,7 @@ final class RomanTypeTest extends TestCase
             'empty labels' => [
                 'sequence' => new Sequence(),
                 'label' => 1,
-                'lettercase' => RomanType::UPPER,
+                'lettercase' => RomanGenerator::UPPER,
                 'expected' => [],
             ],
             'labels starts at 3' => [
@@ -54,19 +54,19 @@ final class RomanTypeTest extends TestCase
                     new Period('2018-02-01', '2018-03-01')
                 ),
                 'label' => 4,
-                'lettercase' => RomanType::UPPER,
+                'lettercase' => RomanGenerator::UPPER,
                 'expected' => ['IV', 'V'],
             ],
             'labels starts at 0 (1)' => [
                 'sequence' => new Sequence(new Period('2018-01-01', '2018-02-01')),
                 'label' => -1,
-                'lettercase' => RomanType::LOWER,
+                'lettercase' => RomanGenerator::LOWER,
                 'expected' => ['i'],
             ],
             'labels starts at 0 (2)' => [
                 'sequence' => new Sequence(new Period('2018-01-01', '2018-02-01')),
                 'label' => 0,
-                'lettercase' => RomanType::LOWER,
+                'lettercase' => RomanGenerator::LOWER,
                 'expected' => ['i'],
             ],
         ];
@@ -74,23 +74,29 @@ final class RomanTypeTest extends TestCase
 
     public function testStartWith(): void
     {
-        $generator = new RomanType(new IntegerType(42));
+        $generator = new RomanGenerator(new IntegerGenerator(42));
         self::assertSame(42, $generator->getStartingAt());
         $new = $generator->startWith(69);
         self::assertNotSame($new, $generator);
         self::assertSame(69, $new->getStartingAt());
         self::assertSame($generator, $generator->startWith(42));
-        self::assertSame(1, (new IntegerType(-3))->getStartingAt());
+        self::assertSame(1, (new IntegerGenerator(-3))->getStartingAt());
         self::assertSame(1, $generator->startWith(-3)->getStartingAt());
     }
 
     public function testLetterCase(): void
     {
-        $generator = new RomanType(new IntegerType(1));
+        $generator = new RomanGenerator(new IntegerGenerator(1));
         self::assertTrue($generator->isUpper());
-        $new = $generator->withLetterCase(RomanType::LOWER);
+        $new = $generator->withLetterCase(RomanGenerator::LOWER);
         self::assertFalse($new->isUpper());
-        $alt = $new->withLetterCase(RomanType::LOWER);
+        $alt = $new->withLetterCase(RomanGenerator::LOWER);
         self::assertSame($alt, $new);
+    }
+
+    public function testFormat(): void
+    {
+        $generator = new RomanGenerator(new IntegerGenerator(10));
+        self::assertSame('', $generator->format([]));
     }
 }