+++ tests/ViewerTest.php
@@ -15,14 +15,15 @@ namespace BakameTest\Period\Visualizer;
 
 use Bakame\Period\Visualizer\ConsoleConfig;
 use Bakame\Period\Visualizer\ConsoleOutput;
-use Bakame\Period\Visualizer\Label\LetterType;
+use Bakame\Period\Visualizer\Label\LetterGenerator;
 use Bakame\Period\Visualizer\Viewer;
+use League\Period\Datepoint;
 use League\Period\Period;
 use League\Period\Sequence;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @coversDefaultClass Bakame\Period\Visualizer\Viewer
+ * @coversDefaultClass \Bakame\Period\Visualizer\Viewer
  */
 final class ViewerTest extends TestCase
 {
@@ -33,21 +34,35 @@ final class ViewerTest extends TestCase
 
     public function setUp(): void
     {
-        $this->view = new Viewer(new ConsoleOutput(new ConsoleConfig()));
+        $this->view = new Viewer();
     }
 
     /**
-     * @covers Bakame\Period\Visualizer\ConsoleOutput
+     * @covers ::__construct
+     */
+    public function testConstructor(): void
+    {
+        $viewer = new Viewer();
+        self::assertInstanceOf(LetterGenerator::class, $viewer->getLabelGenerator());
+        self::assertEquals(new ConsoleOutput(), $viewer->getOutput());
+    }
+
+    /**
      * @covers ::getLabelGenerator
+     * @covers ::setLabelGenerator
      */
     public function testLabelGenerator(): void
     {
         $labelGenerator = $this->view->getLabelGenerator();
-        self::assertInstanceOf(LetterType::class, $labelGenerator);
-        $this->view->setLabelGenerator(new LetterType('bb'));
+        self::assertInstanceOf(LetterGenerator::class, $labelGenerator);
+        $this->view->setLabelGenerator(new LetterGenerator('bb'));
         self::assertNotSame($labelGenerator, $this->view->getLabelGenerator());
     }
 
+    /**
+     * @covers ::getOutput
+     * @covers ::setOutput
+     */
     public function testOutput(): void
     {
         $output = $this->view->getOutput();
@@ -55,6 +70,11 @@ final class ViewerTest extends TestCase
         self::assertNotSame($output, $this->view->getOutput());
     }
 
+    /**
+     * @covers ::sequence
+     * @covers ::filterResultLabel
+     * @covers ::addLabels
+     */
     public function testDisplaySequence(): void
     {
         $data = $this->view->sequence(new Sequence(
@@ -62,16 +82,24 @@ final class ViewerTest extends TestCase
             new Period('2018-01-15', '2018-02-01')
         ));
 
-        self::assertStringContainsString('A     [==================================)', $data);
-        self::assertStringContainsString('B                                         [=========================================)', $data);
+        self::assertStringContainsString('A    [==========================)', $data);
+        self::assertStringContainsString('B                               [===============================)', $data);
     }
 
+    /**
+     * @covers ::sequence
+     */
     public function testDisplayEmptySequence(): void
     {
         $data = $this->view->sequence(new Sequence());
         self::assertEmpty($data);
     }
 
+    /**
+     * @covers ::intersections
+     * @covers ::filterResultLabel
+     * @covers ::addLabels
+     */
     public function testDisplayIntersection(): void
     {
         $data = $this->view->intersections(new Sequence(
@@ -79,11 +107,16 @@ final class ViewerTest extends TestCase
             new Period('2018-01-10', '2018-02-01')
         ));
 
-        self::assertStringContainsString('A                 [==================================)', $data);
-        self::assertStringContainsString('B                                        [======================================================)', $data);
-        self::assertStringContainsString('INTERSECTIONS                            [===========)', $data);
+        self::assertStringContainsString('A                [==========================)', $data);
+        self::assertStringContainsString('B                                 [=========================================)', $data);
+        self::assertStringContainsString('INTERSECTIONS                     [=========)', $data);
     }
 
+    /**
+     * @covers ::gaps
+     * @covers ::filterResultLabel
+     * @covers ::addLabels
+     */
     public function testGaps(): void
     {
         $data = $this->view->gaps(new Sequence(
@@ -91,11 +124,15 @@ final class ViewerTest extends TestCase
             new Period('2018-01-15', '2018-02-01', Period::EXCLUDE_ALL)
         ), '');
 
-        self::assertStringContainsString('A          [=====================)', $data);
-        self::assertStringContainsString('B                                              (=========================================)', $data);
-        self::assertStringContainsString('RESULT                            [===========]', $data);
+        self::assertStringContainsString('A         [================)', $data);
+        self::assertStringContainsString('B                                    (===============================)', $data);
+        self::assertStringContainsString('RESULT                     [=========]', $data);
     }
 
+    /**
+     * @covers ::sequence
+     * @covers ::addLabels
+     */
     public function testSingleUnitIntervalLength(): void
     {
         $data = $this->view->sequence(new Sequence(
@@ -103,25 +140,44 @@ final class ViewerTest extends TestCase
             new Period('2017-01-01', '2019-01-01', Period::INCLUDE_ALL)
         ));
 
-        self::assertStringContainsString('A                                             [=)', $data);
-        self::assertStringContainsString('B     [=============================================================================]', $data);
+        self::assertStringContainsString('A                                  [=)', $data);
+        self::assertStringContainsString('B    [==========================================================]', $data);
     }
 
     /**
-     * @covers Bakame\Period\Visualizer\ConsoleOutput
-     * @covers Bakame\Period\Visualizer\Viewer
+     * @covers ::diff
+     * @covers ::addLabels
      */
     public function testDiff(): void
     {
         $config = (new ConsoleConfig())->withColors('white');
-        $view = new Viewer(new ConsoleOutput($config));
+        $view = new Viewer(new LetterGenerator(), new ConsoleOutput($config));
         $data = $view->diff(
             new Period('2018-01-01', '2018-02-01'),
             new Period('2017-12-01', '2018-03-01')
         );
 
-        self::assertStringContainsString('A                                   [==========================)', $data);
-        self::assertStringContainsString('B        [=============================================================================)', $data);
-        self::assertStringContainsString('DIFF     [=========================)                            [======================)', $data);
+        self::assertStringContainsString('A                           [====================)', $data);
+        self::assertStringContainsString('B       [==========================================================)', $data);
+        self::assertStringContainsString('DIFF    [===================)                    [=================)', $data);
+    }
+
+    /**
+     * @covers ::unions
+     * @covers ::addLabels
+     */
+    public function testUnion(): void
+    {
+        $sequence = new Sequence(
+            Datepoint::create('2018-11-29')->getYear(),
+            Datepoint::create('2018-11-29')->getMonth(),
+            Period::around('2016-06-01', '3 MONTHS')
+        );
+
+        $data = $this->view->unions($sequence);
+        self::assertStringContainsString('A                                               [====================)', $data);
+        self::assertStringContainsString('B                                                                 [=) ', $data);
+        self::assertStringContainsString('C         [=========)', $data);
+        self::assertStringContainsString('UNIONS    [=========)                           [====================)', $data);
     }
 }